Add stub heartbeat (stub_hb) builds
Normal stub builds should behave almost exactly like
normal builds, except that all calls to hardware are hardcoded
with default values. So breaking out the permanent heartbeat feature
into another build seems appropriate.

The normal stub builds are particular useful because we needed
a device that was constantly on to ensure that the gateway saw traffic
on the network. Since we have devices that go down, we needed a control
to compare against to see if it was an error with the gateway or
an error with a physical sensor node.